// MastermindFrontend/src/openapi/model/gameDto.ts
/**
 * MastermindBackend
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Guess } from './guess';


export interface GameDto { 
    id?: string | null;
    user?: string | null;
    maxGuesses?: number | null;
    guesses?: Array<Guess> | null;
}